pyproject.toml
[project]
name = "arxgen"
version = "0.1.0"
description = "Veille scientifique ArXiv avec CrewAI et Ollama"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai[tools]>=0.100.0",
    "arxiv>=2.1.3",
    "langchain-community>=0.3.0",
    "langchain-ollama>=0.2.0",
    "litellm>=1.50.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.1",
    "PyYAML>=6.0.1",
    "fastapi>=0.115.0",
    "starlette>=0.40.0",
    "onnxruntime<1.20.0",
    "markdown>=3.7",
    "weasyprint>=63.0",
]

[project.scripts]
arxgen = "arxgen.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["src/arxgen"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]